Whole cell lysate proteomes of a SILAC experiment of SINV-infected cells during a timecourse experiment. The dataset contains the peptide.txt output file from MaxQuant.

Whole cell lysate proteomes of a SILAC experiment of SINV-infected cells during a timecourse experiment. The dataset contains the peptide.txt output file from MaxQuant.

WCLpeptides.raw

Format

A dataframe with 25136 rows and 148 variables. The important variables are as follows:

Sequence

The identified AA sequence of the peptide.

Length

The length of the sequence stored in the column 'Sequence'.

Proteins

Identifiers of proteins this site is associated with..

“Intensity L ” + M_4_18/18_M_4/4_18_M exp

Summed up eXtracted Ion Current (XIC) of the isotopic cluster linked to the light label partner.

“Intensity M ” + M_4_18/18_M_4/4_18_M exp

Summed up eXtracted Ion Current (XIC) of the isotopic cluster linked to the medium label partner.

“Intensity H ” + M_4_18/18_M_4/4_18_M exp

Summed up eXtracted Ion Current (XIC) of the isotopic cluster linked to the heavy label partner.

Source

Garcia-Moreno et al 2019. System-wide Profiling of RNA-Binding Proteins Uncovers Key Regulators of Virus Infection. Molecular Cell 2019 Apr 4;74(1):196-211.e11 10.1016/j.molcel.2019.01.017.

Value

A data.frame